+++ compras/urls.py
@@ -1,6 +1,7 @@
 from django.urls import path
-from compras.views import compras
+from compras.views import compras, proveedores
 
 urlpatterns = [
-    path('',compras,name="compras"),
+    path('compras/',compras,name="compras"),
+    path('proveedores/', proveedores, name="proveedores"),
 ]
