+++ compras/views.py
@@ -1,9 +1,17 @@
-from multiprocessing import context
 from django.shortcuts import render
 
 # Create your views here.
 
 def compras(request):
+    titulo = "Insumos"
     context={
+        'titulo':titulo
     }
-    return render(request,'compras/compras.html',context)
+    return render(request, "compras/compras.html", context)
+
+def proveedores(request):
+    titulo = "Proveedores"
+    context={
+        'titulo': titulo
+    }
+    return render(request, 'compras/proveedores.html', context)
