+++ principal/settings.py
@@ -38,10 +38,11 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
+    'bootstrap5', # pip install django-bootstrap-v5
     'ventas',
     'usuarios',
     'compras',
-    'bootstrap5', # pip install django-bootstrap-v5
+
 ]
 
 MIDDLEWARE = [
