+++ usuarios/views.py
@@ -4,21 +4,29 @@ from django.shortcuts import render
 # Create your views here.
 
 def usuarios(request):
+    titulo = "Usuarios"
     context={
+        'titulo':titulo
     }
-    return render(request,'usuarios/usuarios.html',context)
-#pensé que podía ser necesario poner el def para cada  ventana aquí, lo dejo por si es necesario en algún momento
-# def insumos(request):
-#     context={
-#     }
-#     return render(request, "usuarios/insumos.html", context)
+    return render(request, "usuarios/usuarios.html", context)
 
-# def servicios(request):
-#     context={
-#     }
-#     return render(request, "usuarios/servicios.html", context)
+def insumos(request):
+    titulo = "Insumos"
+    context={
+        'titulo':titulo
+    }
+    return render(request, "usuarios/insumos.html", context)
 
-# def productos(request):
-#     context={
-#     }
-#     return render(request, "usuarios/productos.html", context)
+def productos(request):
+    titulo = "Productos"
+    context={
+        'titulo':titulo
+    }
+    return render(request, "usuarios/productos.html", context)
+
+def servicios(request):
+    titulo = "Servicios"
+    context={
+        'titulo':titulo
+    }
+    return render(request, "usuarios/servicios.html", context)
