+++ ventas/urls.py
@@ -1,6 +1,7 @@
 from django.urls import path
-from ventas.views import ventas
+from ventas.views import ventas, clientes
 
 urlpatterns = [
-    path('',ventas,name="ventas"),
+    path('ventas/',ventas,name="ventas"),
+    path('clientes/',clientes,name="clientes"),
 ]
